# docs/swagger.yaml
basePath: /
definitions:
  github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      request_id:
        type: string
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta:
    properties:
      cursor:
        type: string
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta'
      success:
        type: boolean
    type: object
...
  internal_opportunity.LegMeta:
    properties:
      best_ask:
        type: string
      best_bid:
        type: string
      health_status:
        type: string
      last_update_age_ms:
        type: integer
      order_book_depth:
        type: integer
      spread:
        type: string
    type: object
  internal_opportunity.LegSide:
    enum:
    - BUY
    - SELL
    type: string
    x-enum-varnames:
    - LegSideBuy
    - LegSideSell
  internal_opportunity.MarketQuality:
    enum:
    - GOOD
    - DEGRADED
    - UNUSABLE
    type: string
    x-enum-varnames:
    - MarketQualityGood
    - MarketQualityDegraded
    - MarketQualityUnusable
  internal_opportunity.Opportunity:
    properties:
      buy_venue_code:
        type: string
      buy_venue_id:
        type: string
      calculation_version:
        type: string
      canonical_symbol:
        type: string
      confidence:
        type: number
      created_at:
        type: string
      detected_at:
        type: string
      estimated_fees:
        type: string
      estimated_other_costs:
        type: string
      estimated_slippage:
        type: string
      expected_holding_secs:
        type: integer
      expected_net_edge:
        type: string
      expected_net_edge_bps:
        type: integer
      expected_net_pnl:
        type: string
      expires_at:
        type: string
      gross_edge:
        type: string
      gross_edge_bps:
        type: integer
      id:
        type: string
      instrument_id:
        type: string
      legs:
        items:
          $ref: '#/definitions/internal_opportunity.OpportunityLeg'
        type: array
      market_quality:
        $ref: '#/definitions/internal_opportunity.MarketQuality'
      market_state_ref:
        type: string
      opportunity_type:
        $ref: '#/definitions/internal_opportunity.OpportunityType'
      sell_venue_code:
        type: string
      sell_venue_id:
        type: string
      status:
        $ref: '#/definitions/internal_opportunity.OpportunityStatus'
      suggested_notional:
        type: string
      suggested_size:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  internal_opportunity.OpportunityLeg:
    properties:
      estimated_slippage_bps:
        type: integer
      expected_funding:
        type: string
      expected_price:
        type: string
      fee_bps:
        type: integer
      id:
        type: string
      instrument_id:
        type: string
      metadata:
        $ref: '#/definitions/internal_opportunity.LegMeta'
      opportunity_id:
        type: string
      side:
        $ref: '#/definitions/internal_opportunity.LegSide'
      target_notional:
        type: string
      target_size:
        type: string
      venue_code:
        type: string
      venue_id:
        type: string
    type: object
  internal_opportunity.OpportunityStatus:
    enum:
    - DETECTED
    - EXECUTED
    - EXPIRED
    - REJECTED
    - CANCELLED
    type: string
    x-enum-varnames:
    - OpportunityStatusDetected
    - OpportunityStatusExecuted
    - OpportunityStatusExpired
    - OpportunityStatusRejected
    - OpportunityStatusCancelled
  internal_opportunity.OpportunityType:
    enum:
    - PRICE_ARBITRAGE
    - FUNDING_ARBITRAGE
    - BASIS_ARBITRAGE
    type: string
    x-enum-varnames:
    - OpportunityTypePriceArb
    - OpportunityTypeFundingArb
    - OpportunityTypeBasisArb
...
host: localhost:8080
info:
  contact:
    email: support@arbitrage.com
    name: API Support
  description: Backend API for arbitrage trading platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Arbitrage Platform API
  version: "1.0"
paths:
...
  /api/v1/opportunities/{id}:
    delete:
      description: Drops one opportunity from the in-memory store; idempotent. Public
        endpoint — no auth.
      parameters:
      - description: Opportunity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: COMMON-902 invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      summary: Remove opportunity (public)
      tags:
      - opportunities
    get:
      description: One opportunity by ID. Public endpoint — no auth.
      parameters:
      - description: Opportunity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_opportunity.Opportunity'
              type: object
        "400":
          description: COMMON-902 invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "404":
          description: COMMON-903 not found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      summary: Get opportunity (public)
      tags:
      - opportunities
  /api/v1/opportunities/instrument/{instrument_id}:
    get:
      description: Opportunities for one instrument ID. Public endpoint — no auth.
      parameters:
      - description: Instrument ID
        in: path
        name: instrument_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_opportunity.Opportunity'
                  type: array
              type: object
        "400":
          description: COMMON-902 invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      summary: List opportunities by instrument (public)
      tags:
      - opportunities
...
  /api/v1/opportunities/type/{type}:
    get:
      description: Opportunities filtered by arbitrage type. Public endpoint — no
        auth.
      parameters:
      - description: Opportunity type
        enum:
        - PRICE_ARBITRAGE
        - FUNDING_ARBITRAGE
        - BASIS_ARBITRAGE
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_opportunity.Opportunity'
                  type: array
              type: object
        "400":
          description: COMMON-902 invalid type
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      summary: List opportunities by type (public)
      tags:
      - opportunities
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"